requirements.txt
torch
einops
#fire
#gradio==5.17.1
#huggingface-hub
#invisible-watermark
#matplotlib
numpy
opencv-python
Pillow
Requests
safetensors
scikit-learn
scipy
scikit-image
torchvision
tqdm
#transformers==4.49.0
sentencepiece